Corregido bug con flag -O, espero que esta vez de verdad...

+++ plantillator/iotools.py
@@ -137,11 +137,8 @@ class ContextMaker(object):
                 self.output_dir = PathElem(os.getcwd())
             else:
                 self.output_dir = PathElem(os.path.dirname(self.outpath))
-                if self.overwrite:
-                    try:
-                        os.unlink(outpath)
-                    except IOError:
-                        pass
+                if self.overwrite and os.path.isfile(outpath):
+                    os.unlink(outpath)
             self.overwrite = False
         else:
             self.output_dir = PathElem(outpath or os.getcwd())
